fix: moved adminBootstrapper in fixtures-build. chore: wrong user error
