+++ src/console/build-fixtures.ts
@@ -12,6 +12,8 @@ import {
   StockDataFixtures,
   VintageDataFixtures,
 } from './fixtures-data/fixtures-models';
+import * as bcrypt from 'bcrypt';
+import { CARBONABLE_SALT } from '../auth/constants';
 
 const ulid = monotonicFactory();
 
@@ -175,10 +177,38 @@ export class BuildFixturesCommand extends CommandRunner {
     );
     return await res.json();
   }
+  async seedUsers() {
+    const data = await this.getAdmin();
+    await this.prisma.user.create({
+      data,
+    });
+  }
+
+  async getAdmin(): Promise<any> {
+    const name = process.env.DEFAULT_ADMIN_NAME;
+    const password = process.env.DEFAULT_ADMIN_PASSWORD;
+    const rolesEnv = process.env.DEFAULT_ADMIN_ROLES;
+
+    if (!rolesEnv || !name || !password) {
+      throw new Error(
+        'DEFAULT_ADMIN_NAME, DEFAULT_ADMIN_PASSWORD, or DEFAULT_ADMIN_ROLES are not defined in the environment variables',
+      );
+    }
+
+    const roles = rolesEnv.replace(/[\[\]']/g, '').split(',');
+    const hashedPassword = await bcrypt.hash(password, CARBONABLE_SALT);
+    return {
+      id: ulid().toString(),
+      name,
+      password: hashedPassword,
+      roles,
+    };
+  }
 
   async run(): Promise<void> {
     try {
       await this.seedCountries();
+      await this.seedUsers();
       await this.addFixtures({
         connection: this.prisma,
       });
