+++ src/main.ts
@@ -2,7 +2,6 @@ import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 import { LogLevel, ValidationPipe } from '@nestjs/common';
 import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
-import { bootstrapAuthAndUsers } from './admin-verification';
 
 const LOG_LEVEL: LogLevel[] =
   process.env.NODE_ENV === 'production'
@@ -18,7 +17,6 @@ async function bootstrap() {
     .setVersion('1.0')
     .addTag('CCPM')
     .build();
-  await bootstrapAuthAndUsers();
   const document = SwaggerModule.createDocument(app, config);
   SwaggerModule.setup('api', app, document);
